refactor(tools): decompose thirteenth-salary page into feature structure
Extracts 1065-LOC monolith into feature folder under
app/features/tools/thirteenth-salary/:
- page.vue — shell/layout component (<250 LOC guideline)
- useThirteenthSalaryCalculator.ts — re-export facade for model functions
- thirteenth-salary.schema.ts — Zod form validation schema
- ThirteenthSalaryResult.vue — props-only result display component
- __tests__/useThirteenthSalaryCalculator.spec.ts — 12 unit tests

Route page app/pages/tools/thirteenth-salary.vue is now a thin wrapper
(SEO meta + layout only, imports feature page.vue).
